Quay login only works if commit is in master
GitHub Action environment variables gets passed to PR only if they
starts from a branch that is local, so fork does not get a QUAY PASSOWRD
and USERNAME, this is to avoid security issue.

GitHub is working at it but for now that's what we have.

We push images to Quay only if the commit is in master, but during a
re-work of the GitHub Action we login all the time. And that's a bug.